more cleaning. Implemented string representation for fitness
