+++ src/bindings/bind_selection.cpp
@@ -5,14 +5,8 @@ namespace py = pybind11;
 namespace br = Brush;
 namespace nl = nlohmann;
 
-// using Reg = br::Program<ArrayXf>;
-// using Cls = br::Program<ArrayXb>;
-// using Rep = br::Program<ArrayXXf>;
-// using MCls = br::Program<ArrayXi>;
-
 void bind_selections(py::module& m)
 {
-    // TODO: make them a single class
     bind_selection<br::ProgramType::Regressor>(m, "RegressorSelector");
     bind_selection<br::ProgramType::BinaryClassifier>(m, "ClassifierSelector");
