+++ src/vary/search_space.cpp
@@ -276,10 +276,6 @@ tree<Node>& SearchSpace::PTC2(tree<Node>& Tree,
     // parameters, the real maximum size that can occur is `max_size` plus the
     // highest operator arity, and the real maximum depth is `max_depth` plus one.
 
-    // auto Tree = tree<Node>();
-
-    // fmt::print("building program with max size {}, max depth {}",max_size,max_d);
-
     // Queue of nodes that need children
     vector<tuple<TreeIter, DataType, int>> queue; 
 
@@ -289,9 +285,6 @@ tree<Node>& SearchSpace::PTC2(tree<Node>& Tree,
     int s = 1;
 
     Node root = spot.node->data;
-
-    // cout << "root " << root.name << endl;
-    // auto spot = Tree.set_head(n);
 
     // updating size accordingly to root node
     if (Is<NodeType::SplitBest>(root.node_type))
@@ -315,8 +308,6 @@ tree<Node>& SearchSpace::PTC2(tree<Node>& Tree,
 
     Node n;
     // Now we actually start the PTC2 procedure to create the program tree
-    // cout << "queue size: " << queue.size() << endl;
-    // cout << "entering first while loop...\n";
     while ( queue.size() + s < max_size && queue.size() > 0) 
     {            
         // including the queue size in the max_size, since each element in queue
@@ -334,12 +325,6 @@ tree<Node>& SearchSpace::PTC2(tree<Node>& Tree,
         // cout << "current depth: " << d << endl;
         if (d >= max_d || s >= max_size)
         {
-            // choose terminal of matching type
-            // cout << "getting " << DataTypeName[t] << " terminal\n";
-            // qspot = sample_terminal(t);
-            // Tree.replace(qspot, sample_terminal(t));
-            // Tree.append_child(qspot, sample_terminal(t));
-
             auto opt = sample_terminal(t);
 
             // if it returned optional, then there's nothing to sample based on weights.
@@ -355,11 +340,7 @@ tree<Node>& SearchSpace::PTC2(tree<Node>& Tree,
         else
         {
             //choose a nonterminal of matching type
-            // cout << "getting op of type " << DataTypeName[t] << endl;
             auto opt = sample_op(t);
-            // cout << "chose " << n.name << endl;
-            // TreeIter new_spot = Tree.append_child(qspot, n);
-            // qspot = n;
 
             if (!opt) { // there is no operator for this node. sample a terminal instead
                 opt = sample_terminal(t);
@@ -380,8 +361,6 @@ tree<Node>& SearchSpace::PTC2(tree<Node>& Tree,
             // For each arg of n, add to queue
             for (auto a : n.arg_types)
             {
-                // cout << "queing a node of type " << DataTypeName[a] << endl;
-                // queue.push_back(make_tuple(new_spot, a, d+1));
                 auto child_spot = Tree.append_child(newspot);
 
                 queue.push_back(make_tuple(child_spot, a, d+1));
@@ -399,24 +378,14 @@ tree<Node>& SearchSpace::PTC2(tree<Node>& Tree,
         if ( n.get_is_weighted()==true
         &&   Isnt<NodeType::Constant, NodeType::MeanLabel>(n.node_type) )
             s += 2;
-
-        // cout << "current tree size: " << s << endl;
     } 
 
-    // cout << "entering second while loop...\n";
     while (queue.size() > 0)
     {
         if (queue.size() == 0)
             break;
 
-        // cout << "queue size: " << queue.size() << endl;
-
         auto [qspot, t, d] = RandomDequeue(queue);
-
-        // cout << "getting " << DataTypeName[t] << " terminal\n";
-        // Tree.append_child(qspot, sample_terminal(t));
-        // qspot = sample_terminal(t);
-        // auto newspot = Tree.replace(qspot, sample_terminal(t));
 
         auto opt = sample_terminal(t);
         if (!opt)
@@ -426,11 +395,6 @@ tree<Node>& SearchSpace::PTC2(tree<Node>& Tree,
         
         auto newspot = Tree.replace(qspot, n);
     }
-
-    // cout << "final tree:\n"
-    //  << Tree.begin().node->get_model() << "\n"
-    //  << Tree.begin().node->get_tree_model(true) << endl;
-
     return Tree;
 };
